split out frame length parsing
